DamageOverTime (Burn) is modified.
No longer affect Fire or Water element from the Fire DoT.
Has its own "damageOverTimeTotalTicks" (In Player Settings) amount to play with how long it's suppose to be active on the target (Just the Enemy for now).

Added a Seed in enemySpawner to get the same results on each playthrough.